提交第一个版本

+++ config/SQL/domain-sample-mysql.sql
@@ -10,7 +10,7 @@ CREATE TABLE family(
 DROP TABLE IF EXISTS family_address;
 CREATE TABLE family_address(
                                `ID` SERIAL(8) NOT NULL AUTO_INCREMENT  COMMENT '自增主键' ,
-                               `family_id` DECIMAL(8)    COMMENT '关联用户' ,
+                               `family_id` BIGINT    COMMENT '关联用户' ,
                                `address_name` VARCHAR(255)    COMMENT '地址' ,
                                PRIMARY KEY (ID)
 )  COMMENT = '用户家庭住址';
@@ -18,7 +18,7 @@ CREATE TABLE family_address(
 DROP TABLE IF EXISTS family_member;
 CREATE TABLE family_member(
                               `ID` SERIAL(8) NOT NULL AUTO_INCREMENT  COMMENT '自增主键' ,
-                              `family_id` DECIMAL(8)    COMMENT '家庭ID' ,
+                              `family_id` BIGINT    COMMENT '家庭ID' ,
                               `name` VARCHAR(255)    COMMENT '姓名' ,
                               `phone` VARCHAR(255)    COMMENT '电话' ,
                               `type` VARCHAR(255)    COMMENT '成员关系' ,
